evict outgoing sdk tests from awslamda
